Add files via upload
A College Voting System is a web-based application designed to facilitate the voting process in a college environment, allowing students to cast their votes online in a secure and convenient manner. The system is developed using HTML, CSS, JavaScript, PHP, and MySQL to create a user-friendly interface, handle backend processing, and store data in a database.